move maven and gradle cleanup to modular
