+++ generators/gradle/cleanup.mts
