+++ generators/maven/cleanup.mts
