+++ generators/maven/generator.mjs
@@ -19,12 +19,12 @@
 /* eslint-disable consistent-return */
 import assert from 'assert/strict';
 
-import { threadId } from 'worker_threads';
 import BaseApplicationGenerator from '../base-application/index.mjs';
 
 import { GENERATOR_MAVEN, GENERATOR_BOOTSTRAP_APPLICATION_SERVER } from '../generator-list.mjs';
 import files from './files.mjs';
 import { MAVEN } from './constants.mjs';
+import cleanupOldServerFilesTask from './cleanup.mjs';
 
 /**
  * @class
@@ -62,6 +62,7 @@ export default class MavenGenerator extends BaseApplicationGenerator {
 
   get writing() {
     return this.asWritingTaskGroup({
+      cleanupOldServerFilesTask,
       async writeFiles() {
         await this.writeFiles({ sections: files, context: this.application });
       },
